Implement num_elements_wrong_position function
Function was intended to count the number of code elements that are the
correct value, but wrong position. However, during testing it became
clear that the current implementation is flawwed.

The current implementation works through elements of a guess once, from
first to last. However, logic is needed to account for cases where a
guess such as ["C, "C", "C", "C"] is evaluated against a target ["A",
"C", "A", "A"]. In the current implementation, the first element is
evaluated as being a correct value but in the wrong place. However, the
intended behaviour is to realise that the second element is the correct
value in the correct place, which overrides the former assessment.